reactive state per thread
